langchain-ollama/requirements.txt
langchain_community
langchain
python-dotenv
faiss-cpu
pypdf
requests
twilio
google-search-results
wolframalpha
fastmcp
ollama
openai
